Refactor shared helpers into utility modules
